requirements.txt
Django>=3.2
pandas>=2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pyarrow>=12.0
tqdm>=4.66
requests>=2.31
beautifulsoup4>=4.12
joblib>=1.3
scikit-learn>=1.4
